+++ lib/ruby_aem/resources/flush_agent.rb
@@ -13,6 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 =end
+require 'uri'
 
 module RubyAem
   module Resources
@@ -50,6 +51,10 @@ module RubyAem
         @call_params[:title] = title
         @call_params[:description] = description
         @call_params[:dest_base_url] = dest_base_url
+
+        uri = URI.parse(dest_base_url)
+        @call_params[:ssl] = uri.scheme == 'https' ? 'relaxed' : ''
+
         @call_params = @call_params.merge(opts)
         @client.call(self.class, __callee__.to_s, @call_params)
       end