+++ test/unit/resources/flush_agent_spec.rb
@@ -21,6 +21,7 @@ describe 'FlushAgent' do
           :title => 'Some Flush Agent Title',
           :description => 'Some flush agent description',
           :dest_base_url => 'http://somehost:8080',
+          :ssl => '',
           :log_level => 'error',
           :retry_delay => 30000 })
       @flush_agent.create_update('Some Flush Agent Title', 'Some flush agent description', 'http://somehost:8080')
@@ -38,7 +39,7 @@ describe 'FlushAgent' do
           :ssl => 'relaxed',
           :log_level => 'info',
           :retry_delay => 60000 })
-      @flush_agent.create_update('Some Flush Agent Title', 'Some flush agent description', 'https://somehost:8080', { log_level: 'info', retry_delay: 60000, ssl: 'relaxed' })
+      @flush_agent.create_update('Some Flush Agent Title', 'Some flush agent description', 'https://somehost:8080', { log_level: 'info', retry_delay: 60000 })
     end
 
   end